split big Emitter to submodules.
